# apps/fauxify-api-client/models/album.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fauxify API
 * API for managing artists, albums, songs, and analytics.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Song } from './song';

/**
 * 
 * @export
 * @interface Album
 */
export interface Album {
    /**
     * 
     * @type {string}
     * @memberof Album
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Album
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof Album
     */
    'artistId'?: string;
    /**
     * 
     * @type {string}
     * @memberof Album
     */
    'releaseDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof Album
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof Album
     */
    'updatedAt'?: string;
    /**
     * 
     * @type {Array<Song>}
     * @memberof Album
     */
    'songs'?: Array<Song>;
}