# apps/fauxify-api-client/models/song.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fauxify API
 * API for managing artists, albums, songs, and analytics.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Song
 */
export interface Song {
    /**
     * 
     * @type {string}
     * @memberof Song
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Song
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof Song
     */
    'albumId'?: string;
    /**
     * 
     * @type {number}
     * @memberof Song
     */
    'trackNumber'?: number;
    /**
     * 
     * @type {number}
     * @memberof Song
     */
    'durationSeconds'?: number;
    /**
     * 
     * @type {string}
     * @memberof Song
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof Song
     */
    'updatedAt'?: string;
}